Fix file taint for linux guests.
A previous commit introduced get_ntreadfile_esp_off() for
working around an issue with file taint for Windows 2000
guests.  This function can only be called when the
wintrospection plugin is loaded.  Wintrospection is not loaded
for linux guests.

This change ensures the function is only called when the
guest is a windows os.